refactor: move all paradigm files to js/paradigm.js
